+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.2
+
+- remove `create_copied_hierarchy_without_top_most_dir` and now use the unified quick pick to decide what to create
+- Added quick pick dialog when creating folder hierarchy to allow users to select which path segments to include